requirements.txt
#vbcable not a python library
pyttsx3
pyaudio
torch
torchaudio
TTS
gradio
SpeechRecognition
psutil
pystray
pillow
wmi
requests
wikipedia-api
wikipedia
sumy
sounddevice
librosa # maybe not
scipy
vosk # no need
pystray
pillow